requirements.txt
psycopg2-binary==2.9.9
table2ascii
discord.py==2.3.2
pillow==9.0.1
